requirements.txt
django
djangorestframework
django-jazzmin
drf-spectacular
pillow
django-filter
requests